Makefile created for project

+++ PY2C++/Scanner.cpp
@@ -231,7 +231,7 @@ void Scanner :: scanToken(){
 }
 
 bool Scanner :: isAtEnd() {
-	return current >= source.length();
+	return current >= (int)source.length();
 }
